requirements.txt
pip install python-jose[cryptography]
pip install bcrypt==4.0.1
python-dotenv